+++ packages/map-template/src/components/MapTemplate/MapTemplate.jsx
@@ -55,6 +55,7 @@ import currentVenueNameState from '../../atoms/currentVenueNameState.js';
 import categoryState from '../../atoms/categoryState.js';
 import hideNonMatchesState from '../../atoms/hideNonMatchesState.js';
 import showExternalIDsState from '../../atoms/showExternalIDsState.js'
+import showRoadNamesState from '../../atoms/showRoadNamesState.js';
 
 // Define the Custom Elements from our components package.
 defineCustomElements();
@@ -89,7 +90,7 @@ defineCustomElements();
  * @param {boolean} [props.hideNonMatches] - Determine whether the locations on the map should be filtered (only show the matched locations and hide the rest) or highlighted (show all locations and highlight the matched ones with a red dot by default). If set to true, the locations will be filtered.
  * @param {boolean} [props.showExternalIDs] - Determine whether the location details on the map should have an external ID visible. The default value is set to false.
  */
-function MapTemplate({ apiKey, gmApiKey, mapboxAccessToken, venue, locationId, primaryColor, logo, appUserRoles, directionsFrom, directionsTo, externalIDs, tileStyle, startZoomLevel, bearing, pitch, gmMapId, useMapProviderModule, kioskOriginLocationId, language, supportsUrlParameters, useKeyboard, timeout, miTransitionLevel, category, searchAllVenues, hideNonMatches, showExternalIDs }) {
+function MapTemplate({ apiKey, gmApiKey, mapboxAccessToken, venue, locationId, primaryColor, logo, appUserRoles, directionsFrom, directionsTo, externalIDs, tileStyle, startZoomLevel, bearing, pitch, gmMapId, useMapProviderModule, kioskOriginLocationId, language, supportsUrlParameters, useKeyboard, timeout, miTransitionLevel, category, searchAllVenues, hideNonMatches, showRoadNames, showExternalIDs }) {
 
     const [, setApiKey] = useRecoilState(apiKeyState);
     const [, setGmApiKey] = useRecoilState(gmApiKeyState);
@@ -117,6 +118,7 @@ function MapTemplate({ apiKey, gmApiKey, mapboxAccessToken, venue, locationId, p
     const [, setCategory] = useRecoilState(categoryState);
     const [, setHideNonMatches] = useRecoilState(hideNonMatchesState);
     const [, setshowExternalIDs] = useRecoilState(showExternalIDsState);
+    const [, setShowRoadNames] = useRecoilState(showRoadNamesState);
 
     const [showVenueSelector, setShowVenueSelector] = useState(true);
     const [showPositionControl, setShowPositionControl] = useState(true);
@@ -213,7 +215,8 @@ function MapTemplate({ apiKey, gmApiKey, mapboxAccessToken, venue, locationId, p
             const languageToUse = language ? language : navigator.language;
 
             // Set the language on the MapsIndoors SDK in order to get eg. Mapbox and Google directions in that language.
-            window.mapsindoors.MapsIndoors.setLanguage(languageToUse);
+            // The MapsIndoors data only accepts the first part of the IETF language string, hence the split.
+            window.mapsindoors.MapsIndoors.setLanguage(languageToUse.split('-')[0]);
 
             // If relevant, fetch venues, categories and the current location again to get them in the new language
             window.mapsindoors.services.LocationsService.once('update_completed', () => {
@@ -243,6 +246,7 @@ function MapTemplate({ apiKey, gmApiKey, mapboxAccessToken, venue, locationId, p
                 // Change the already set language
                 i18n.changeLanguage(languageToUse);
             }
+
 
             setCurrentLanguage(languageToUse);
         }
@@ -541,6 +545,13 @@ function MapTemplate({ apiKey, gmApiKey, mapboxAccessToken, venue, locationId, p
         }
     }, [showExternalIDs]);
 
+    /*
+     * React on changes to the showRoadNames prop.
+     */
+    useEffect(() => {
+        setShowRoadNames(showRoadNames);
+    }, [showRoadNames])
+
     /**
      * When venue is fitted while initializing the data,
      * set map to be ready and get the venue categories.
@@ -594,7 +605,6 @@ function MapTemplate({ apiKey, gmApiKey, mapboxAccessToken, venue, locationId, p
      * @param {array} locationsResult
      */
     function getCategories(locationsResult) {
-        // Initialise the unique categories map
         let uniqueCategories = new Map();
 
         // Loop through the locations and count the unique locations.
@@ -606,19 +616,20 @@ function MapTemplate({ apiKey, gmApiKey, mapboxAccessToken, venue, locationId, p
                 // Get the categories from the App Config that have a matching key.
                 const appConfigCategory = appConfig?.menuInfo.mainmenu.find(category => category.categoryKey === key);
 
-                if (uniqueCategories.has(key)) {
-                    let count = uniqueCategories.get(key).count;
-                    uniqueCategories.set(key, { count: ++count, displayName: location.properties.categories[key], iconUrl: appConfigCategory?.iconUrl });
-                } else {
-                    uniqueCategories.set(key, { count: 1, displayName: location.properties.categories[key], iconUrl: appConfigCategory?.iconUrl });
+                if (appConfigCategory) {
+                    uniqueCategories.set(appConfigCategory.categoryKey, { displayName: location.properties.categories[key], iconUrl: appConfigCategory?.iconUrl })
                 }
             }
         }
 
-        // Sort the categories with most locations associated.
-        uniqueCategories = Array.from(uniqueCategories).sort((a, b) => b[1].count - a[1].count);
+        // Sort categories by the place in the mainmenu array. Use index to do that.
+        const sortedCategories = Array.from(uniqueCategories).sort((a, b) => {
+            const orderA = appConfig.menuInfo.mainmenu.findIndex(category => category.categoryKey === a[0]);
+            const orderB = appConfig.menuInfo.mainmenu.findIndex(category => category.categoryKey === b[0]);
+            return orderA - orderB;
+        });
 
-        setCategories(uniqueCategories);
+        setCategories(sortedCategories);
     }
 
     /*